plugins/discord/requirements.txt
discord.py[voice]>=2.3.0
aiohttp>=3.9.0
PyYAML>=6.0
PyNaCl>=1.5.0
pytest-asyncio>=0.23.0
